refactor: promote group_vars_by_dims to a shared df.py utility
Both the DataFusion path and the DuckDB adapter split mixed-dimension
Datasets with it; living private in sql.py made the generic backends
package import the DataFusion-specific module. Tests move with it.
