+++ tests/test_df.py
@@ -16,6 +16,7 @@ from xarray_sql.df import (
     explode,
     from_map,
     from_map_batched,
+    group_vars_by_dims,
     iter_record_batches,
     partition_metadata,
     pivot,
@@ -716,3 +717,53 @@ def test_partition_metadata_in_range_datetime_still_pruned():
     for m in meta:
         _, _, tag = m["time"]
         assert tag == "timestamp_ns"
+
+
+class TestGroupVarsByDims:
+    def test_single_dim_group(self):
+        ds = xr.Dataset(
+            {
+                "a": (["x", "y"], np.zeros((2, 3))),
+                "b": (["x", "y"], np.ones((2, 3))),
+            }
+        )
+        groups = group_vars_by_dims(ds)
+        assert groups == {("x", "y"): ["a", "b"]}
+
+    def test_multiple_dim_groups(self):
+        ds = xr.Dataset(
+            {
+                "surface": (["time", "lat", "lon"], np.zeros((2, 3, 4))),
+                "upper": (
+                    ["time", "lat", "lon", "level"],
+                    np.zeros((2, 3, 4, 5)),
+                ),
+            }
+        )
+        groups = group_vars_by_dims(ds)
+        assert set(groups.keys()) == {
+            ("time", "lat", "lon"),
+            ("time", "lat", "lon", "level"),
+        }
+        assert groups[("time", "lat", "lon")] == ["surface"]
+        assert groups[("time", "lat", "lon", "level")] == ["upper"]
+
+    def test_empty_dataset(self):
+        assert group_vars_by_dims(xr.Dataset()) == {}
+
+    def test_includes_scalar_group(self):
+        """Scalar (0-dim) variables group under the empty dims tuple."""
+        ds = xr.Dataset(
+            {"band": (["y", "x"], np.zeros((2, 3))), "projection": ((), 0)}
+        )
+        groups = group_vars_by_dims(ds)
+        assert groups == {("y", "x"): ["band"], (): ["projection"]}
+
+    def test_ignores_coords(self):
+        """Coordinate variables shouldn't be returned as groups."""
+        ds = xr.Dataset(
+            {"v": (["x"], np.arange(3))},
+            coords={"x": np.arange(3), "label": ("x", ["a", "b", "c"])},
+        )
+        groups = group_vars_by_dims(ds)
+        assert groups == {("x",): ["v"]}
